Remove partition rereading tests for reverted fixes
The change that block/013 tested for was reverted in Linux kernel commit
10c70d95c0f2 ("block: remove the bd_openers checks in
blk_drop_partitions"). To quote Christoph:

"That check only catches file systems that use a single block device
(e.g. not btrfs multi-device or XFS or ext4 with log devices) and also
doesn't catch non-filesystem users. I first tried to generalized it,
but that ran into a chain of other problems. And there really isn't
much of a problem re-reading partitions on a mounted file system - it is
pointless but not actually harmful."

So, we shouldn't expect that check to come back. Let's remove the
test.

Similarly, the change that scsi/003 tested for was reverted in Linux
kernel commit 8acf608e602f ("Revert "scsi: sd: Keep disk read-only when
re-reading partition""). According to that commit, this can be fixed, so
when that happens we can reintroduce the test.

Reported-by: Yi Zhang <yi.zhang@redhat.com>
Signed-off-by: Omar Sandoval <osandov@fb.com>